+++ unlaxer-common/src/main/java/org/unlaxer/parser/NonTerminalSymbol.java
@@ -1,0 +1,7 @@
+package org.unlaxer.parser;
+
+import java.io.Serializable;
+
+/** Marker for parser symbols that are composed from other grammar elements. */
+public interface NonTerminalSymbol extends Serializable {
+}